+++ src/assets/scripts/client/InputController.js
@@ -1,10 +1,11 @@
 /* eslint-disable camelcase, no-mixed-operators, object-shorthand, class-methods-use-this, no-undef, expected-return*/
 import $ from 'jquery';
-import _get from 'lodash/get';
 import _has from 'lodash/has';
 import _map from 'lodash/map';
 import CommandParser from './commandParser/CommandParser';
+import EventBus from './lib/EventBus';
 import { clamp } from './math/core';
+import { EVENT } from './constants/eventNames';
 import { GAME_OPTION_NAMES } from './constants/gameOptionConstants';
 import { SELECTORS } from './constants/selectors';
 
@@ -84,8 +85,13 @@ const KEY_CODES = {
 export default class InputController {
     /**
      * @constructor
+     * @param $element {JQuery|HTML Element}
+     * @param aircraftCommander {AircraftCommander}
+     * @param uiController {UiController}
+     * @param onSelectAircraftStrip {function}       provides direct access to method in AircraftController that
+     *                                               that can be used to select a specific `stripViewModel`.
      */
-    constructor($element, aircraftCommander) {
+    constructor($element, aircraftCommander, uiController, aircraftController) {
         this.$element = $element;
         this.$window = null;
         this.$commandInput = null;
@@ -93,7 +99,12 @@ export default class InputController {
         this.$sidebar = null;
 
         this._aircraftCommander = aircraftCommander;
+        this._uiController = uiController;
+        this._aircraftController = aircraftController;
 
+        this._eventBus = EventBus;
+
+        prop.input = input;
         this.input = input;
         this.input.command = '';
         this.input.callsign = '';
@@ -130,6 +141,7 @@ export default class InputController {
      * @method setupHandlers
      */
     setupHandlers() {
+
         return this;
     }
 
@@ -150,6 +162,8 @@ export default class InputController {
         this.$canvases.on('mouseup', (event) => this.onMouseUpHandler(event));
         this.$canvases.on('mousedown', (event) => this.onMouseDownHandler(event));
 
+        this._eventBus.on(EVENT.STRIP_CLICK, this.input_select);
+
         return this;
     }
 
@@ -166,6 +180,8 @@ export default class InputController {
         this.$canvases.off('mousemove', (event) => this.onMouseMoveHandler(event));
         this.$canvases.off('mouseup', (event) => this.onMouseUpHandler(event));
         this.$canvases.off('mousedown', (event) => this.onMouseDownHandler(event));
+
+        this._eventBus.off(EVENT.STRIP_CLICK, this.input_select);
 
         return this.destroy();
     }
@@ -202,19 +218,19 @@ export default class InputController {
      * @method input_init_pre
      */
     input_init_pre() {
-        // TODO: these prop properties can be removed except for `prop.input`
-        prop.input = input;
-        prop.input.command = '';
-        prop.input.callsign = '';
-        prop.input.data = '';
-        prop.input.history = [];
-        prop.input.history_item = null;
-        prop.input.click = [0, 0];
-        prop.input.positions = '';
-        prop.input.tab_compl = {};
-        prop.input.mouseDelta = [0, 0];
-        prop.input.mouseDown = [0, 0];
-        prop.input.isMouseDown = false;
+        // TODO: these prop properties can be removed except for `this.input`
+        this.input = input;
+        this.input.command = '';
+        this.input.callsign = '';
+        this.input.data = '';
+        this.input.history = [];
+        this.input.history_item = null;
+        this.input.click = [0, 0];
+        this.input.positions = '';
+        this.input.tab_compl = {};
+        this.input.mouseDelta = [0, 0];
+        this.input.mouseDown = [0, 0];
+        this.input.isMouseDown = false;
     }
 
     /**
@@ -224,9 +240,9 @@ export default class InputController {
      */
     onMouseScrollHandler(event) {
         if (event.originalEvent.wheelDelta > 0 || event.originalEvent.detail < 0) {
-            window.uiController.ui_zoom_in();
+            this._uiController.ui_zoom_in();
         } else {
-            window.uiController.ui_zoom_out();
+            this._uiController.ui_zoom_out();
         }
     }
 
@@ -236,16 +252,16 @@ export default class InputController {
      * @param event {jquery Event}
      */
     onMouseMoveHandler(event) {
-        if (!prop.input.isMouseDown) {
+        if (!this.input.isMouseDown) {
             return this;
         }
 
-        prop.input.mouseDelta = [
-            event.pageX - prop.input.mouseDown[0],
-            event.pageY - prop.input.mouseDown[1]
+        this.input.mouseDelta = [
+            event.pageX - this.input.mouseDown[0],
+            event.pageY - this.input.mouseDown[1]
         ];
-        prop.canvas.panX = prop.input.mouseDelta[0];
-        prop.canvas.panY = prop.input.mouseDelta[1];
+        prop.canvas.panX = this.input.mouseDelta[0];
+        prop.canvas.panY = this.input.mouseDelta[1];
         prop.canvas.dirty = true;
     }
 
@@ -255,7 +271,7 @@ export default class InputController {
      * @param event {jquery Event}
      */
     onMouseUpHandler(event) {
-        prop.input.isMouseDown = false;
+        this.input.isMouseDown = false;
     }
 
     /**
@@ -266,44 +282,47 @@ export default class InputController {
     onMouseDownHandler(event) {
         event.preventDefault();
 
+        // TODO: this should use early returns instead of the else if
         if (event.which === MOUSE_EVENT_CODE.MIDDLE_PESS) {
-            window.uiController.ui_zoom_reset();
+            this._uiController.ui_zoom_reset();
         } else if (event.which === MOUSE_EVENT_CODE.LEFT_PRESS) {
             // Record mouse down position for panning
-            prop.input.mouseDown = [
+            this.input.mouseDown = [
                 event.pageX - prop.canvas.panX,
                 event.pageY - prop.canvas.panY
             ];
-            prop.input.isMouseDown = true;
+            this.input.isMouseDown = true;
 
             // Aircraft label selection
             let position = [event.pageX, -event.pageY];
             position[0] -= prop.canvas.size.width / 2;
             position[1] += prop.canvas.size.height / 2;
 
-            const nearest = window.aircraftController.aircraft_get_nearest([
-                window.uiController.px_to_km(position[0] - prop.canvas.panX),
-                window.uiController.px_to_km(position[1] + prop.canvas.panY)
+            const [aircraftModel, distanceFromPosition] = this._aircraftController.aircraft_get_nearest([
+                this._uiController.px_to_km(position[0] - prop.canvas.panX),
+                this._uiController.px_to_km(position[1] + prop.canvas.panY)
             ]);
 
-            if (nearest[0]) {
-                if (nearest[1] < window.uiController.px_to_km(80)) {
-                    this.input.callsign = nearest[0].callsign.toUpperCase();
+            if (aircraftModel) {
+                if (distanceFromPosition < this._uiController.px_to_km(80)) {
+                    this.input.callsign = aircraftModel.callsign.toUpperCase();
 
                     this.input_select(this.input.callsign);
+                    this._eventBus.trigger(EVENT.SELECT_STRIP_VIEW_FROM_DATA_BLOCK, aircraftModel);
                 } else {
                     this.input_select();
+                    this._eventBus.trigger(EVENT.DESELECT_ACTIVE_STRIP_VIEW, {});
                 }
             }
 
             position = [
-                window.uiController.px_to_km(position[0]),
-                window.uiController.px_to_km(position[1])
+                this._uiController.px_to_km(position[0]),
+                this._uiController.px_to_km(position[1])
             ];
 
             position[0] = parseFloat(position[0].toFixed(2));
             position[1] = parseFloat(position[1].toFixed(2));
-            prop.input.positions += `[${position.join(',')}]`;
+            this.input.positions += `[${position.join(',')}]`;
 
             return false;
         }
@@ -327,19 +346,19 @@ export default class InputController {
             if (prop.tutorial.open) {
                 window.tutorialView.tutorial_close();
             } else if ($(SELECTORS.DOM_SELECTORS.AIRPORT_SWITCH).hasClass(SELECTORS.CLASSNAMES.OPEN)) {
-                window.uiController.ui_airport_close();
+                this._uiController.ui_airport_close();
             }
         }
 
         if (event.which === KEY_CODES.DASH || (is_firefox && event.which === KEY_CODES.DASH_FIREFOX)) {
             // Minus key to zoom out, plus to zoom in
-            window.uiController.ui_zoom_out();
+            this._uiController.ui_zoom_out();
             return false;
         } else if (event.which === KEY_CODES.EQUALS || (is_firefox && event.which === KEY_CODES.EQUALS_FIREFOX)) {
             if (event.shiftKey) {
-                window.uiController.ui_zoom_in();
+                this._uiController.ui_zoom_in();
             } else {
-                window.uiController.ui_zoom_reset();
+                this._uiController.ui_zoom_reset();
             }
 
             return false;
@@ -363,50 +382,32 @@ export default class InputController {
      * @method input_parse
      */
     input_parse() {
-        const $strip = this.$element.find(SELECTORS.DOM_SELECTORS.STRIP);
-        $strip.removeClass(SELECTORS.CLASSNAMES.ACTIVE);
+        this.input.callsign = '';
+        this.input.data = '';
 
-        prop.input.callsign = '';
-        prop.input.data = '';
-
-        if (prop.input.command.length === 0) {
+        if (this.input.command.length === 0) {
             return;
         }
 
         // TODO: move to master REGEX constant
-        let match = /^\s*(\w+)/.exec(prop.input.command);
+        const match = /^\s*(\w+)/.exec(this.input.command);
 
         if (!match) {
             return;
         }
 
-        prop.input.callsign = match[1];
-        let number = 0;
-        // TODO: this is a very mutable property. perhaps it should be something else?
-        match = null;
+        this.input.callsign = match[1];
         prop.canvas.dirty = true;
 
-        for (let i = 0; i < prop.aircraft.list.length; i++) {
-            const aircraft = prop.aircraft.list[i];
+        // TODO: this looks like it should happen in the `AircraftController`
+        for (let i = 0; i < this._aircraftController.aircraft.list.length; i++) {
+            const aircraft = this._aircraftController.aircraft.list[i];
 
-            if (aircraft.matchCallsign(prop.input.callsign)) {
-                number += 1;
-                match = aircraft;
-                // TODO: this should be from an encapsulated class on the window.
-                aircraft.$html.addClass(SELECTORS.CLASSNAMES.ACTIVE);
+            if (aircraft.matchCallsign(this.input.callsign)) {
+                this._aircraftController.onSelectAircraftStrip(aircraft);
+
+                break;
             }
-        }
-
-        // TODO: this logic block should be either abstracted or simplified.
-        if (number === 1 && (
-                match.$html.offset().top < 0 ||
-                (
-                    (match.$html.offset().top + match.$html.height() - this.$sidebar.offset().top) >
-                    this.$sidebar.height()
-                )
-            )
-        ) {
-            this.$sidebar.scrollTop(this.$sidebar.scrollTop() + match.$html.offset().top - (this.$sidebar.height() / 2));
         }
     }
 
@@ -417,7 +418,7 @@ export default class InputController {
     onCommandInputChangeHandler() {
         this.tab_completion_reset();
 
-        prop.input.command = this.$commandInput.val();
+        this.input.command = this.$commandInput.val();
 
         this.input_parse();
     }
@@ -427,17 +428,18 @@ export default class InputController {
      * @method input_select
      * @param callsign {string}
      */
-    input_select(callsign) {
+    input_select = (callsign) => {
+        let nextCommandInputValue = '';
+
         if (callsign) {
-            this.$commandInput.val(`${callsign} `);
-        } else {
-            this.$commandInput.val('');
+            nextCommandInputValue = `${callsign} `;
         }
 
+        this.$commandInput.val(nextCommandInputValue);
         this.$commandInput.focus();
 
         this.onCommandInputChangeHandler();
-    }
+    };
 
     /**
      * @for InputController
@@ -458,30 +460,29 @@ export default class InputController {
                 this.input_parse();
 
                 if (this.input_run()) {
-                    prop.input.history.unshift(prop.input.callsign);
+                    this.input.history.unshift(this.input.callsign);
                     this.$commandInput.val('');
-                    prop.input.command = '';
+                    this.input.command = '';
 
                     this.tab_completion_reset();
                     this.input_parse();
                 }
 
-                prop.input.history_item = null;
+                this.input.history_item = null;
 
                 break;
-
             case KEY_CODES.PAGE_UP:
                 // recall previous callsign
                 this.input_history_prev();
                 e.preventDefault();
-                break;
 
+                break;
             case KEY_CODES.PAGE_DOWN:
                 // recall subsequent callsign
                 this.input_history_next();
                 e.preventDefault();
-                break;
 
+                break;
             case KEY_CODES.LEFT_ARROW:
                 // shortKeys in use
                 if (this._isArrowControlMethod()) {
@@ -491,7 +492,6 @@ export default class InputController {
                 }
 
                 break;
-
             case KEY_CODES.UP_ARROW:
                 if (this._isArrowControlMethod()) {
                     this.$commandInput.val(`${currentCommandInputValue} \u2B61 `);
@@ -502,8 +502,8 @@ export default class InputController {
                     this.input_history_prev();
                     e.preventDefault();
                 }
-                break;
 
+                break;
             case KEY_CODES.RIGHT_ARROW:
                 // shortKeys in use
                 if (this._isArrowControlMethod()) {
@@ -513,7 +513,6 @@ export default class InputController {
                 }
 
                 break;
-
             case KEY_CODES.DOWN_ARROW:
                 if (this._isArrowControlMethod()) {
                     this.$commandInput.val(`${currentCommandInputValue} \u2B63 `);
@@ -526,51 +525,44 @@ export default class InputController {
                 }
 
                 break;
-
             case KEY_CODES.MULTIPLY:
                 this.$commandInput.val(`${currentCommandInputValue} \u2B50 `);
                 e.preventDefault();
                 this.onCommandInputChangeHandler();
 
                 break;
-
             case KEY_CODES.ADD:
                 this.$commandInput.val(`${currentCommandInputValue} + `);
                 e.preventDefault();
                 this.onCommandInputChangeHandler();
 
                 break;
-
             case KEY_CODES.EQUALS: // mac + (actually `=`)
                 this.$commandInput.val(`${currentCommandInputValue} + `);
                 e.preventDefault();
                 this.onCommandInputChangeHandler();
 
                 break;
-
             case KEY_CODES.SUBTRACT:
                 this.$commandInput.val(`${currentCommandInputValue} - `);
                 e.preventDefault();
                 this.onCommandInputChangeHandler();
 
                 break;
-
             case KEY_CODES.DASH: // mac -
                 this.$commandInput.val(`${currentCommandInputValue} - `);
                 e.preventDefault();
                 this.onCommandInputChangeHandler();
 
                 break;
-
             case KEY_CODES.DIVIDE:
                 this.$commandInput.val(`${currentCommandInputValue} takeoff `);
                 e.preventDefault();
                 this.onCommandInputChangeHandler();
 
                 break;
-
             case KEY_CODES.TAB:
-                if (!prop.input.tab_compl.matches) {
+                if (!this.input.tab_compl.matches) {
                     this.tab_completion_match();
                 }
 
@@ -578,7 +570,6 @@ export default class InputController {
                 e.preventDefault();
 
                 break;
-
             case KEY_CODES.ESCAPE:
                 const currentCommandValue = this.$commandInput.val();
 
@@ -603,14 +594,14 @@ export default class InputController {
      * @param opt
      */
     tab_completion_cycle(opt) {
-        const matches = prop.input.tab_compl.matches;
+        const matches = this.input.tab_compl.matches;
 
         if (!matches || matches.length === 0) {
             return;
         }
 
         // TODO: this block needs some work. this initial assignment looks to be overwritten every time.
-        let i = prop.input.tab_compl.cycle_item;
+        let i = this.input.tab_compl.cycle_item;
         if (opt.backwards) {
             i = (i <= 0) ? matches.length - 1 : i - 1;
         } else {
@@ -619,8 +610,8 @@ export default class InputController {
 
         this.$commandInput.val(`${matches[i]} `);
 
-        prop.input.command = matches[i];
-        prop.input.tab_compl.cycle_item = i;
+        this.input.command = matches[i];
+        this.input.tab_compl.cycle_item = i;
 
         this.input_parse();
     }
@@ -632,11 +623,11 @@ export default class InputController {
     tab_completion_match() {
         let matches;
         const val = this.$commandInput.val();
-        let aircrafts = prop.aircraft.list;
+        let aircrafts = this._aircraftController.aircraft.list;
 
-        if (prop.input.callsign) {
+        if (this.input.callsign) {
             aircrafts = aircrafts.filter((a) => {
-                return a.matchCallsign(prop.input.callsign);
+                return a.matchCallsign(this.input.callsign);
             });
         }
 
@@ -644,20 +635,20 @@ export default class InputController {
             return aircraft.callsign;
         });
 
-        if (aircrafts.length === 1 && (prop.input.data || val[val.length - 1] === ' ')) {
+        if (aircrafts.length === 1 && (this.input.data || val[val.length - 1] === ' ')) {
             // TODO: update inline functions
             matches = aircrafts[0].COMMANDS.filter((c) => {
-                return c.toLowerCase().indexOf(prop.input.data.toLowerCase()) === 0;
+                return c.toLowerCase().indexOf(this.input.data.toLowerCase()) === 0;
             })
             .map((c) => {
-                return val.substring(0, prop.input.callsign.length + 1) + c;
+                return val.substring(0, this.input.callsign.length + 1) + c;
             });
         }
 
         this.tab_completion_reset();
 
-        prop.input.tab_compl.matches = matches;
-        prop.input.tab_compl.cycle_item = -1;
+        this.input.tab_compl.matches = matches;
+        this.input.tab_compl.cycle_item = -1;
     }
 
     /**
@@ -665,7 +656,7 @@ export default class InputController {
      * @method tab_completion_reset
      */
     tab_completion_reset() {
-        prop.input.tab_compl = {};
+        this.input.tab_compl = {};
     }
 
     /**
@@ -673,7 +664,7 @@ export default class InputController {
      * @method input_history_clamp
      */
     input_history_clamp() {
-        prop.input.history_item = clamp(0, prop.input.history_item, prop.input.history.length - 1);
+        this.input.history_item = clamp(0, this.input.history_item, this.input.history.length - 1);
     }
 
     /**
@@ -681,19 +672,19 @@ export default class InputController {
      * @method input_history_prev
      */
     input_history_prev() {
-        if (prop.input.history.length === 0) {
+        if (this.input.history.length === 0) {
             return;
         }
 
-        if (prop.input.history_item == null) {
-            prop.input.history.unshift(prop.input.command);
-            prop.input.history_item = 0;
+        if (this.input.history_item == null) {
+            this.input.history.unshift(this.input.command);
+            this.input.history_item = 0;
         }
 
-        prop.input.history_item += 1;
+        this.input.history_item += 1;
         this.input_history_clamp();
 
-        const command = `${prop.input.history[prop.input.history_item]} `;
+        const command = `${this.input.history[this.input.history_item]} `;
         this.$commandInput.val(command.toUpperCase());
 
         this.onCommandInputChangeHandler();
@@ -704,26 +695,26 @@ export default class InputController {
      * @method input_history_next
      */
     input_history_next() {
-        if (prop.input.history.length === 0 || !prop.input.history_item) {
+        if (this.input.history.length === 0 || !this.input.history_item) {
             return;
         }
 
-        prop.input.history_item -= 1;
+        this.input.history_item -= 1;
 
-        if (prop.input.history_item <= 0) {
-            this.$commandInput.val(prop.input.history[0]);
+        if (this.input.history_item <= 0) {
+            this.$commandInput.val(this.input.history[0]);
 
             this.onCommandInputChangeHandler();
 
-            prop.input.history.splice(0, 1);
-            prop.input.history_item = null;
+            this.input.history.splice(0, 1);
+            this.input.history_item = null;
 
             return;
         }
 
         this.input_history_clamp();
 
-        const command = `${prop.input.history[prop.input.history_item]} `;
+        const command = `${this.input.history[this.input.history_item]} `;
 
         this.$commandInput.val(command.toUpperCase());
         this.onCommandInputChangeHandler();
@@ -748,7 +739,7 @@ export default class InputController {
     _parseUserCommand() {
         let result;
         // this could use $commandInput.val() as an alternative
-        const userCommand = prop.input.command.trim().toLowerCase();
+        const userCommand = this.input.command.trim().toLowerCase();
 
         // Using try/catch here very much on purpose. the `CommandParser` will throw when it encounters any kind
         // of error; invalid length, validation, parse, etc. Here we catch those errors, log them to the screen
@@ -756,7 +747,7 @@ export default class InputController {
         try {
             result = new CommandParser(userCommand);
         } catch (error) {
-            window.uiController.ui_log('Command not understood');
+            this._uiController.ui_log('Command not understood');
 
             throw error;
         }
@@ -787,7 +778,7 @@ export default class InputController {
     processSystemCommand(commandParser) {
         switch (commandParser.command) {
             case PARSED_COMMAND_NAME.VERSION:
-                window.uiController.ui_log(`Air Traffic Control simulator version ${prop.version}`);
+                this._uiController.ui_log(`Air Traffic Control simulator version ${prop.version}`);
 
                 return true;
 
@@ -800,10 +791,10 @@ export default class InputController {
                 // TODO: does this function exist anywhere?
                 // aircraft_toggle_auto();
                 //
-                // if (prop.aircraft.auto.enabled) {
-                //     window.uiController.ui_log('automatic controller ENGAGED');
+                // if (this._aircraftController.aircraft.auto.enabled) {
+                //     this._uiController.ui_log('automatic controller ENGAGED');
                 // } else {
-                //     window.uiController.ui_log('automatic controller OFF');
+                //     this._uiController.ui_log('automatic controller OFF');
                 // }
 
                 return true;
@@ -826,6 +817,7 @@ export default class InputController {
                 localStorage.clear();
                 location.reload();
 
+                break;
             case PARSED_COMMAND_NAME.AIRPORT:
                 // TODO: it may be better to do this in the parser
                 const airportIcao = commandParser.args[0];
@@ -859,8 +851,8 @@ export default class InputController {
         let matches = 0;
         let match = -1;
 
-        for (let i = 0; i < prop.aircraft.list.length; i++) {
-            const aircraft = prop.aircraft.list[i];
+        for (let i = 0; i < this._aircraftController.aircraft.list.length; i++) {
+            const aircraft = this._aircraftController.aircraft.list[i];
 
             if (aircraft.matchCallsign(commandParser.callsign)) {
                 matches += 1;
@@ -869,18 +861,18 @@ export default class InputController {
         }
 
         if (matches > 1) {
-            window.uiController.ui_log('multiple aircraft match the callsign, say again');
+            this._uiController.ui_log('multiple aircraft match the callsign, say again');
 
             return true;
         }
 
         if (match === -1) {
-            window.uiController.ui_log('no such aircraft, say again');
+            this._uiController.ui_log('no such aircraft, say again');
 
             return true;
         }
 
-        const aircraft = prop.aircraft.list[match];
+        const aircraft = this._aircraftController.aircraft.list[match];
 
         return this._aircraftCommander.runCommands(aircraft, commandParser.args);
     }