+++ src/assets/scripts/client/aircraft/AircraftInstanceModel.js
@@ -1,12 +1,11 @@
 /* eslint-disable max-len, no-undef */
-import $ from 'jquery';
 import _defaultTo from 'lodash/defaultTo';
 import _forEach from 'lodash/forEach';
 import _get from 'lodash/get';
 import _isEqual from 'lodash/isEqual';
 import _isNil from 'lodash/isNil';
+import _random from 'lodash/random';
 import _uniqueId from 'lodash/uniqueId';
-import AircraftStripView from './AircraftStripView';
 import Fms from './FlightManagementSystem/Fms';
 import ModeController from './ModeControl/ModeController';
 import Pilot from './Pilot/Pilot';
@@ -42,20 +41,17 @@ import {
 } from '../utilities/radioUtilities';
 import {
     degreesToRadians,
-    heading_to_string,
     nm
 } from '../utilities/unitConverters';
 import {
     FLIGHT_CATEGORY,
     FLIGHT_PHASE,
-    PERFORMANCE,
-    WAYPOINT_NAV_MODE
+    PERFORMANCE
 } from '../constants/aircraftConstants';
 import {
     AIRPORT_CONSTANTS,
     AIRPORT_CONTROL_POSITION_NAME
 } from '../constants/airportConstants';
-import { SELECTORS } from '../constants/selectors';
 import { GAME_EVENTS } from '../game/GameController';
 import { MCP_MODE, MCP_MODE_NAME } from './ModeControl/modeControlConstants';
 import { TIME } from '../constants/globalConstants';
@@ -91,7 +87,8 @@ export default class AircraftInstanceModel {
         this.airlineId      = '';         // Airline Identifier (eg. 'AAL')
         this.airlineCallsign = '';
         this.flightNumber = '';         // Flight Number ONLY (eg. '551')
-        this.transponderCode = '';
+        // FIXME: temporary inclusion, should be removed before merging to develop
+        this.transponderCode = _random(1201, 5999);
         this.heading      = 0;          // Magnetic Heading
         this.altitude     = 0;          // Altitude, ft MSL
         this.speed        = 0;          // Indicated Airspeed (IAS), knots
@@ -102,14 +99,35 @@ export default class AircraftInstanceModel {
         this.approachDistance = 0;      // Distance longitudinally from the threshold
         this.radial       = 0;          // Angle from airport center to aircraft
         this.distance     = 0;          //
-        this.destination  = null;       // Destination they're flying to
+
+        /**
+         * The origin ariport for an aircraft
+         *
+         * This will only be populated for dpearture aircraft
+         *
+         * @property origin
+         * @type {string}
+         * @default ''
+         */
+        this.origin = '';
+
+        /**
+         * The destination airpot of an aircraft
+         *
+         * This will only be populated for arrivals
+         *
+         * @property destination
+         * @type {string}
+         * @default ''
+         */
+        this.destination = '';
+
         this.trend        = 0;          // Indicator of descent/level/climb (1, 0, or 1)
         this.history      = [];         // Array of previous positions
         this.restricted   = { list: [] };
         this.notice       = false;      // Whether aircraft
         this.warning      = false;      //
         this.hit          = false;      // Whether aircraft has crashed
-        this.taxi_next    = false;      //
         this.taxi_start   = 0;          //
         this.taxi_time    = 3;          // Time spent taxiing to the runway. *NOTE* this should be INCREASED to around 60 once the taxi vs LUAW issue is resolved (#406)
         this.rules        = FLIGHT_RULES.IFR;      // Either IFR or VFR (Instrument/Visual Flight Rules)
@@ -145,10 +163,6 @@ export default class AircraftInstanceModel {
          */
         this.isRemovable = false;
 
-        // TODO: change name, and update refs in `InputController`. perhaps change to be a ref to the AircraftStripView class instead of directly accessing the html?
-        this.aircraftStripView = null;
-        this.$html = null;
-        this.$strips = $(SELECTORS.DOM_SELECTORS.STRIPS);
         /* eslint-enable multi-spaces*/
 
         // Set to true when simulating future movements of the aircraft
@@ -184,7 +198,6 @@ export default class AircraftInstanceModel {
             ? window.gameController.game_time()
             : null;
 
-
         this.buildCurrentTerrainRanges();
         this.buildRestrictedAreaLinks();
         this.parse(options);
@@ -203,9 +216,6 @@ export default class AircraftInstanceModel {
 
             this.mcp.initializeForAirborneFlight(bottomAltitude, this.heading, this.speed);
         }
-
-        this.createStrip();
-        this.updateStrip();
     }
 
     /**
@@ -283,7 +293,7 @@ export default class AircraftInstanceModel {
     }
 
     parse(data) {
-        // TODO: these _gets can likely be removed
+        // TODO: the `_get()` below are doing the same thing as a `_defaultTo()` would do
         this.positionModel = _get(data, 'positionModel', this.positionModel);
         this.model = _get(data, 'model', this.model);
         this.airlineId = _get(data, 'airline', this.airlineId);
@@ -293,14 +303,16 @@ export default class AircraftInstanceModel {
         this.heading = _get(data, 'heading', this.heading);
         this.altitude = _get(data, 'altitude', this.altitude);
         this.speed = _get(data, 'speed', this.speed);
+        this.origin = _get(data, 'origin', this.origin);
         this.destination = _get(data, 'destination', this.destination);
+
+        // TODO: this assumes and arrival spawns outside the airspace
         this.inside_ctr = data.category === FLIGHT_CATEGORY.DEPARTURE;
     }
 
     initFms(data) {
         const airport = window.airportController.airport_get();
         const initialRunway = airport.getActiveRunwayForCategory(this.category);
-
         this.fms = new Fms(data, initialRunway, this.model, this._navigationLibrary);
 
         if (this.category === FLIGHT_CATEGORY.DEPARTURE) {
@@ -319,6 +331,45 @@ export default class AircraftInstanceModel {
     }
 
     /**
+     * Build an object that contains all the correct data, in the correct shape,
+     * so it can be injected into the view.
+     *
+     * This method should only be used by the `StripView` classes when instantiating
+     * or updating the aircraft progress strips.
+     *
+     * The data here should be considered read-only.
+     *
+     * @for AircraftModel
+     * @method getViewModel
+     * @return {object<string, string>}
+     */
+    getViewModel() {
+        let assignedAltitude = this.mcp.altitude;
+        let flightPlanAltitude = this.fms.flightPlanAltitude;
+
+        if (assignedAltitude === -1) {
+            assignedAltitude = '-';
+        }
+
+        if (flightPlanAltitude === -1) {
+            flightPlanAltitude = '-';
+        }
+
+        return {
+            id: this.id,
+            insideCenter: this.inside_ctr,
+            callsign: this.callsign,
+            transponderCode: this.transponderCode,
+            icaoWithWeightClass: this.model.icaoWithWeightClass,
+            assignedAltitude,
+            flightPlanAltitude,
+            arrivalAirportId: this.destination.toUpperCase(),
+            departureAirportId: this.origin.toUpperCase(),
+            flightPlan: this.fms.getFlightPlanRouteForStripView()
+        };
+    }
+
+    /**
      * Called when the aircraft crosses the airspace boundary (ie, leaving our airspace)
      *
      * @for AircraftInstanceModel
@@ -334,7 +385,6 @@ export default class AircraftInstanceModel {
 
         // Crossing into the center
         if (this.inside_ctr) {
-            this.showStrip();
             this.callUp();
 
             return;
@@ -353,7 +403,6 @@ export default class AircraftInstanceModel {
             return this.arrivalExit();
         }
 
-        this.hideStrip();
         this.setIsRemovable();
 
         // TODO: this seems redundant. if its already in the leg its in the fms.
@@ -503,6 +552,14 @@ export default class AircraftInstanceModel {
      */
     isAirborne() {
         return !this.isOnGround();
+    }
+
+    isDeparture() {
+        return this.fms.isDeparture();
+    }
+
+    isArrival() {
+        return this.fms.isArrival();
     }
 
     /**
@@ -1269,7 +1326,6 @@ export default class AircraftInstanceModel {
     updateLandingFailedLanding() {
         // Failed Approach
         if ((this.approachDistance > 0.100) && (!this.projected)) {
-            this.updateStrip();
             this.cancelLanding();
 
             const isWarning = true;
@@ -1810,8 +1866,9 @@ export default class AircraftInstanceModel {
                 }
             }
 
+            // this was a $.each() and may need to verified that its working with _forEach()
             // raise warning if in at least one restricted area
-            $.each(this.restricted.list, (k, v) => {
+            _forEach(this.restricted.list, (k, v) => {
                 warning = warning || v.inside;
             });
         }
@@ -1903,7 +1960,6 @@ export default class AircraftInstanceModel {
         this.updateFlightPhase();
         this.updateTarget();
         this.updatePhysics();
-        this.updateStrip();
     }
 
     /**
@@ -1955,152 +2011,5 @@ export default class AircraftInstanceModel {
      */
     removeConflict(conflictingAircraft) {
         delete this.conflicts[conflictingAircraft.callsign];
-    }
-
-    // TODO: aircraft strip methods below will be abstracted and de-coupled from this model
-
-    /**
-     * Create the aircraft's flight strip and add to strip bay
-     */
-    createStrip() {
-        this.aircraftStripView = new AircraftStripView(this);
-
-        this.$html = this.aircraftStripView.$element;
-        // Add the strip to the html
-        const scrollPos = this.$strips.scrollTop();
-        this.$strips.prepend(this.aircraftStripView.$element);
-        // shift scroll down one strip's height
-        this.$strips.scrollTop(scrollPos + this.aircraftStripView.height);
-
-        // Determine whether or not to show the strip in our bay
-        if (this.category === FLIGHT_CATEGORY.ARRIVAL) {
-            this.aircraftStripView.hide();
-        }
-    }
-
-    // TODO: move these view methods to `AircraftStripView` or a different file
-    /**
-     * @for AircraftInstanceModel
-     * @method updateStrip
-     */
-    updateStrip() {
-        if (this.projected || !this.inside_ctr) {
-            return;
-        }
-
-        const heading = heading_to_string(this.mcp.heading);
-        const altitude = this.mcp.altitude;
-        const speed = this.mcp.speed;
-        let destinationDisplay = this.fms.getProcedureAndExitName();
-        const altitudeText = this.taxi_next
-            ? 'ready'
-            : null;
-        const hasAltitude = this.mcp.altitude !== -1;
-
-        this.aircraftStripView.update(heading, altitude, this.destination, speed);
-
-        switch (this.flightPhase) {
-            case FLIGHT_PHASE.APRON:
-                this.aircraftStripView.updateViewForApron(destinationDisplay, hasAltitude);
-
-                break;
-            case FLIGHT_PHASE.TAXI:
-                this.aircraftStripView.updateViewForTaxi(destinationDisplay, hasAltitude, altitudeText);
-
-                break;
-            case FLIGHT_PHASE.WAITING:
-                this.aircraftStripView.updateViewForWaiting(
-                    destinationDisplay,
-                    this.pilot.hasDepartureClearance,
-                    hasAltitude
-                );
-
-                break;
-            case FLIGHT_PHASE.TAKEOFF:
-                this.aircraftStripView.updateViewForTakeoff(destinationDisplay);
-
-                break;
-            case FLIGHT_PHASE.CLIMB:
-            case FLIGHT_PHASE.HOLD:
-            case FLIGHT_PHASE.DESCENT:
-            case FLIGHT_PHASE.CRUISE:
-                let cruiseNavMode = WAYPOINT_NAV_MODE.FIX;
-                let headingDisplay = this.fms.currentWaypoint.name.toUpperCase();
-                const isFollowingSid = this.fms.isFollowingSid();
-                const isFollowingStar = this.fms.isFollowingStar();
-                const fixRestrictions = {
-                    altitude: this.fms.currentWaypoint.altitudeRestriction !== -1,
-                    speed: this.fms.currentWaypoint.speedRestriction !== -1
-                };
-
-                // TODO: this will need to be addressed when the AircraftStripView is refactored
-                // this block is a bandaid to prevent `destinationDisplay` from being undefined
-                destinationDisplay = this.fms.currentWaypoint.name;
-
-                if (isFollowingStar || isFollowingSid) {
-                    destinationDisplay = this.fms.getProcedureAndExitName();
-                }
-
-                if (this.fms.currentWaypoint.isHold) {
-                    cruiseNavMode = WAYPOINT_NAV_MODE.HOLD;
-                    headingDisplay = 'holding';
-                    destinationDisplay = this.fms.getDestinationName();
-                } else if (this.mcp.headingMode === MCP_MODE.HEADING.HOLD) {
-                    headingDisplay = this.mcp.headingInDegrees;
-                    destinationDisplay = this.fms.getDestinationName();
-                } else if (this.mcp.headingMode === MCP_MODE.HEADING.VOR_LOC) {
-                    cruiseNavMode = WAYPOINT_NAV_MODE.RWY;
-                    headingDisplay = 'intercept';
-                    destinationDisplay = this.fms.getDestinationAndRunwayName();
-                }
-
-                this.aircraftStripView.updateViewForCruise(
-                    cruiseNavMode,
-                    headingDisplay,
-                    destinationDisplay,
-                    isFollowingSid,
-                    isFollowingStar,
-                    fixRestrictions
-                );
-
-                break;
-            case FLIGHT_PHASE.APPROACH:
-            case FLIGHT_PHASE.LANDING:
-                destinationDisplay = this.fms.getDestinationAndRunwayName();
-
-                this.aircraftStripView.updateViewForLanding(destinationDisplay);
-
-                break;
-            default:
-                throw new TypeError(`Invalid FLIGHT_MODE ${this.flightPhase} passed to .updateStrip()`);
-        }
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method showStrip
-     */
-    showStrip() {
-        this.$html.detach();
-
-        const scrollPos = this.$strips.scrollTop();
-
-        this.$strips.prepend(this.$html);
-        this.$html.show();
-        // TODO enumerate the magic number
-        // shift scroll down one strip's height
-        this.$strips.scrollTop(scrollPos + 45);
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method hideStrip
-     */
-    hideStrip() {
-        this.$html.hide(600);
-    }
-
-    cleanup() {
-        this.$html.remove();
     }
 }